Moved get_choices from views.py to __init__ GET parameter for filtering is called "field" instead of "table" Added inline editing to the Django admin url is vocabularies
