+++ djvocab/serializers.py
@@ -5,6 +5,7 @@ Vocabulary serializers
 from rest_framework import serializers, viewsets
 from djvocab import models
 
+
 class VocabularySerializer(serializers.ModelSerializer):
     '''
     Serializer for vocabulary
@@ -19,6 +20,7 @@ class VocabularySerializer(serializers.ModelSerializer):
             'order'
         )
 
+
 class VocabularyViewSet(viewsets.ModelViewSet):
     '''
     ViewSet for Vocabulary models, allows filtering on list
@@ -27,7 +29,7 @@ class VocabularyViewSet(viewsets.ModelViewSet):
 
     def get_queryset(self):
         queryset = models.Vocabulary.objects.all()
-        table = self.request.query_params.get('table', None)
-        if table:
-            queryset = queryset.filter(key=table)
+        field = self.request.query_params.get('field', None)
+        if field:
+            queryset = queryset.filter(key=field)
         return queryset
